Move secret implementation to secretstore package.
